Tests/expat: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@44851 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ test/expat/elements.c
@@ -56,7 +56,7 @@ main(int argc, char *argv[])
       fprintf(stderr,
               "%s at line %" XML_FMT_INT_MOD "u\n",
               XML_ErrorString(XML_GetErrorCode(parser)),
-              XML_GetCurrentLineNumber(parser));
+              (unsigned long)XML_GetCurrentLineNumber(parser));
       return 1;
     }
   } while (!done);
